server/requirements.txt
fastmcp>=3.0
anthropic>=0.90
python-dotenv
uvicorn
requests
